requirements.txt
protobuf
langchain-google-genai
pillow
python-dotenv
SpeechRecognition
pyttsx3
PyAudio
langchain-openai
pandas
streamlit
langchain
langchain-community
langchainhub
langchain-chroma
bs4
pypdf
langchain-google-vertexai
PyPDF2
chromadb
faiss-cpu
chainlit
langchain_openai
gTTS